Иконки на рабочем столе и контекстное меню

+++ src/main/kotlin/models/IconObjectDesktop.kt
@@ -4,6 +4,6 @@ data class IconObjectDesktop(
     val name: String,
     val path: String,
     val iconPath: String,
-    val x: Int,
-    val y: Int
+    val x: Float,
+    val y: Float
 )
